break down query operators
