+++ src/maggma/api/query_operator/__init__.py
@@ -1,0 +1,4 @@
+from maggma.api.query_operator.core import QueryOperator
+from maggma.api.query_operator.pagination import PaginationQuery
+from maggma.api.query_operator.sparse_fields import SparseFieldsQuery
+from maggma.api.query_operator.dynamic import DefaultDynamicQuery
