+++ src/maggma/api/query_operator/dynamic.py
@@ -7,163 +7,10 @@ from monty.json import MSONable
 from pydantic import BaseModel
 from pydantic.fields import ModelField
 
-from maggma.api.util import STORE_PARAMS, dynamic_import
+from maggma.api.util import STORE_PARAMS
 from maggma.core import Store
-
-
-class QueryOperator(MSONable):
-    """
-    Base Query Operator class for defining powerfull query language
-    in the Materials API
-    """
-
-    def query(self) -> STORE_PARAMS:
-        """
-        The query function that does the work for this query operator
-        """
-        raise NotImplementedError("Query operators must implement query")
-
-    def meta(self, store: Store, query: Dict) -> Dict:
-        """
-        Returns meta data to return with the Response
-
-        Args:
-            store: the Maggma Store that the resource uses
-            query: the query being executed in this API call
-        """
-        return {}
-
-    def post_process(self, doc: Dict) -> Dict:
-        """
-        An optional post-processing function for the data
-        """
-        return doc
-
-
-class PaginationQuery(QueryOperator):
-    """Query opertators to provides Pagination in the Materials API"""
-
-    def __init__(
-        self, default_skip: int = 0, default_limit: int = 10, max_limit: int = 100
-    ):
-        """
-        Args:
-            default_skip: the default number of documents to skip
-            default_limit: the default number of documents to return
-            max_limit: max number of documents to return
-        """
-        self.default_skip = default_skip
-        self.default_limit = default_limit
-        self.max_limit = max_limit
-
-        def query(
-            skip: int = Query(
-                default_skip, description="Number of entries to skip in the search"
-            ),
-            limit: int = Query(
-                default_limit,
-                description="Max number of entries to return in a single query."
-                f" Limited to {max_limit}",
-            ),
-        ) -> STORE_PARAMS:
-            """
-            Pagination parameters for the API Endpoint
-            """
-            if limit > max_limit:
-                raise Exception(
-                    "Requested more data per query than allowed by this endpoint."
-                    f"The max limit is {max_limit} entries"
-                )
-            return {"skip": skip, "limit": limit}
-
-        self.query = query  # type: ignore
-
-    def meta(self, store: Store, query: Dict) -> Dict:
-        """
-        Metadata for the pagination params
-        """
-        return {"max_limit": self.max_limit}
-
-
-class SparseFieldsQuery(QueryOperator):
-    def __init__(self, model: BaseModel, default_fields: Optional[List[str]] = None):
-        """
-        Args:
-            model: PyDantic Model that represents the underlying data source
-            default_fields: default fields to return in the API response if no fields are explicitly requested
-        """
-
-        self.model = model
-
-        model_fields = list(self.model.__fields__.keys())
-        self.default_fields = (
-            model_fields if default_fields is None else list(default_fields)
-        )
-
-        assert set(self.default_fields).issubset(
-            model_fields
-        ), "default projection contains some fields that are not in the model fields"
-
-        default_fields_string = ",".join(self.default_fields)  # type: ignore
-
-        def query(
-            fields: str = Query(
-                default_fields_string,
-                description=f"Fields to project from {model.__name__} "  # type: ignore
-                f"as a list of comma seperated strings",
-            ),
-            all_fields: bool = Query(False, description="Include all fields."),
-        ) -> STORE_PARAMS:
-            """
-            Projection parameters for the API Endpoint
-            """
-
-            projection_fields: List[str] = [s.strip() for s in fields.split(",")]
-
-            # need to strip to avoid input such as "name, weight" to be parsed into ["name", " weight"]
-            # we need ["name", "weight"]
-            if all_fields:
-                projection_fields = model_fields
-            return {"properties": projection_fields}
-
-        self.query = query  # type: ignore
-
-    """
-    Factory function to generate a dependency for sparse field sets in FastAPI
-    """
-
-    def meta(self, store: Store, query: Dict) -> Dict:
-        """
-        Returns metadata for the Sparse field set
-        """
-        return {"default_fields": self.default_fields}
-
-    def as_dict(self) -> Dict:
-        """
-        Special as_dict implemented to convert pydantic models into strings
-        """
-
-        d = super().as_dict()  # Ensures sub-classes serialize correctly
-        d["model"] = f"{self.model.__module__}.{self.model.__name__}"  # type: ignore
-        return d
-
-    @classmethod
-    def from_dict(cls, d):
-        """
-        Special from_dict to autoload the pydantic model from the location string
-        """
-        model = d.get("model")
-        if isinstance(model, str):
-            module_path = ".".join(model.split(".")[:-1])
-            class_name = model.split(".")[-1]
-            model = dynamic_import(module_path, class_name)
-
-        assert issubclass(
-            model, BaseModel
-        ), "The resource model has to be a PyDantic Model"
-        d["model"] = model
-
-        cls(**d)
+from maggma.utils import dynamic_import
+from maggma.api.query_operator import QueryOperator
 
 
 class DefaultDynamicQuery(QueryOperator):
